Move navigation buttons to header
